added eslint and fixed eslint errors
