+++ .eslintrc
@@ -1,0 +1,31 @@
+
+{
+  "env": {
+    "browser": true,
+    "es6": true,
+    "node": true,
+    "jest": true
+  },
+  "extends": ["airbnb"],
+  "parser": "babel-eslint",
+  "parserOptions": {
+    "ecmaFeatures": {
+      "experimentalObjectRestSpread": true,
+      "jsx": true
+    },
+    "sourceType": "module"
+  },
+  "rules": {
+    "no-console": "off",
+    "max-len": [1, 110, 2, { "ignoreComments": true }],
+    "jsx-a11y/no-autofocus": [
+      1,
+      {
+        "ignoreNonDOM": true
+      }
+    ],
+    "react/jsx-uses-vars": 1,
+    "indent": [2, 2, {"SwitchCase": 1}]
+  },
+  "plugins": ["react", "jsx-a11y", "import"]
+}
